Consolidate SymCrypt X.509 parsing
