+++ crypto/src/ecdsa_signature.rs
@@ -1,6 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+use x509_cert::der::{asn1::UintRef, Reader, SliceReader};
+
 use super::{EcSignatureKeyAlgorithm, Result};
 
 pub(crate) fn fixed_from_components(
@@ -30,119 +32,43 @@ pub(crate) fn ecdsa_der_to_fixed(
     signature: &[u8],
     algorithm: EcSignatureKeyAlgorithm,
 ) -> Result<Vec<u8>> {
-    let mut index = 0;
-    if signature.get(index) != Some(&0x30) {
-        return Err("ECDSA signature must be a DER SEQUENCE".into());
-    }
-    index += 1;
-
-    let sequence_len = read_der_len(signature, &mut index)?;
-    let sequence_end = index
-        .checked_add(sequence_len)
-        .ok_or("ECDSA signature DER length overflow")?;
-    if sequence_end != signature.len() {
-        return Err("ECDSA signature DER SEQUENCE length does not match input".into());
-    }
-
-    let r = read_der_integer_fixed(signature, &mut index, sequence_end, "r", algorithm)?;
-    let s = read_der_integer_fixed(signature, &mut index, sequence_end, "s", algorithm)?;
-    if index != sequence_end {
-        return Err("ECDSA signature DER SEQUENCE has trailing data".into());
-    }
+    let mut reader = SliceReader::new(signature)?;
+    let components = reader.sequence(|reader| {
+        Ok((
+            reader.decode::<UintRef<'_>>()?,
+            reader.decode::<UintRef<'_>>()?,
+        ))
+    });
+    let (r, s) = reader
+        .finish(components?)
+        .map_err(|e| format!("Failed to parse DER ECDSA signature: {e:?}"))?;
 
     let mut fixed = Vec::with_capacity(algorithm.fixed_signature_byte_len());
-    fixed.extend_from_slice(&r);
-    fixed.extend_from_slice(&s);
+    append_fixed_width_integer(&mut fixed, r.as_bytes(), "r", algorithm)?;
+    append_fixed_width_integer(&mut fixed, s.as_bytes(), "s", algorithm)?;
     Ok(fixed)
 }
 
-fn read_der_len(input: &[u8], index: &mut usize) -> Result<usize> {
-    let first = *input
-        .get(*index)
-        .ok_or("Unexpected end of DER while reading length")?;
-    *index += 1;
-
-    if first & 0x80 == 0 {
-        return Ok(first as usize);
-    }
-
-    let len_len = (first & 0x7f) as usize;
-    if len_len == 0 {
-        return Err("Indefinite DER lengths are not supported".into());
-    }
-    if len_len > std::mem::size_of::<usize>() {
-        return Err("DER length is too large".into());
-    }
-    if input.len().saturating_sub(*index) < len_len {
-        return Err("Unexpected end of DER while reading long-form length".into());
-    }
-    if input[*index] == 0 {
-        return Err("DER length has a redundant leading zero".into());
-    }
-
-    let mut len = 0usize;
-    for byte in &input[*index..*index + len_len] {
-        len = len
-            .checked_mul(256)
-            .and_then(|len| len.checked_add(*byte as usize))
-            .ok_or("DER length overflow")?;
-    }
-    *index += len_len;
-    if len < 0x80 {
-        return Err("DER length uses non-minimal long-form encoding".into());
-    }
-    Ok(len)
-}
-
-fn read_der_integer_fixed(
-    input: &[u8],
-    index: &mut usize,
-    limit: usize,
+fn append_fixed_width_integer(
+    output: &mut Vec<u8>,
+    integer: &[u8],
     name: &str,
     algorithm: EcSignatureKeyAlgorithm,
-) -> Result<Vec<u8>> {
-    if *index >= limit || input.get(*index) != Some(&0x02) {
-        return Err(format!("ECDSA signature DER missing INTEGER {name}").into());
-    }
-    *index += 1;
-
-    let len = read_der_len(input, index)?;
-    let end = index
-        .checked_add(len)
-        .ok_or("ECDSA signature DER INTEGER length overflow")?;
-    if len == 0 || end > limit {
-        return Err(format!("Invalid ECDSA signature DER INTEGER {name} length").into());
-    }
-
-    let mut component = &input[*index..end];
-    *index = end;
-
-    if component[0] & 0x80 != 0 {
-        return Err(format!("ECDSA signature DER INTEGER {name} is negative").into());
-    }
-    if component.len() > 1 && component[0] == 0 {
-        if component[1] & 0x80 == 0 {
-            return Err(
-                format!("ECDSA signature DER INTEGER {name} has redundant sign padding").into(),
-            );
-        }
-        component = &component[1..];
-    }
-
+) -> Result<()> {
     let expected_len = algorithm.scalar_byte_len();
-    if component.len() > expected_len {
+    if integer.len() > expected_len {
         return Err(format!(
             "Invalid ECDSA {} {name} component length: expected <= {}, got {}",
             algorithm.name(),
             expected_len,
-            component.len()
+            integer.len()
         )
         .into());
     }
 
-    let mut fixed = vec![0; expected_len];
-    fixed[expected_len - component.len()..].copy_from_slice(component);
-    Ok(fixed)
+    output.resize(output.len() + expected_len - integer.len(), 0);
+    output.extend_from_slice(integer);
+    Ok(())
 }
 
 #[cfg(test)]
